Make Cassandra healthcheck optional.
Having the healthcheck query cassandra was proven to be a bad idea. If
cassandra is overloaded, that'd cause the healthcheck to fail. That in
turns causes the client HAProxy to run out of healthy backends to route
requests to and return 503 errors.

It's still useful to be able to healthcheck cassandra for itests, since
we don't want to start running tests until Cassandra is ready. So I made
checking cassandra optional. We only check it if `check_cassandra=true`
is set in the HTTP request.