Run integration tests against Cassandra 3.11.3.
We're planning to upgrade our prod Cassandra cluster to 3.11.1. Let's
make sure the cassandra driver and our code work fine with the new C*
version by running the itests against Cassandra 3.11.

I also updated the tests to run on Python 3 since Python 2 required an
extra library `scandir` which doesn't have wheels on the public pypi and
was failing to compile.