Build image again in itests
Jenkins changes the DOCKER_TAG variable, which makes cook-image create
an image with a different name.
Right now docker-compose expects the image to be called
spectre-dev-$USER, which isn't true anymore. The easiest way to fix this
is to rebuild the image during itests. All the layers should be cached
so it's gonna be very fast.