Support routing requests though envoy
Right now we assume that we use haproxy / smartstack to reach the
destination service. This change adds the possibility to route the
requests through envoy instead.

The only thing we really need to change is the function that returns the
ip / port combination for the destination service. Plus add the 'Host'
header that envoy uses to decide what service to route to.