Don't mutate the request headers
When using envoy to route outgoing requests, we need to add the correct
Host header. However if we mutate the request headers this will cause
the cache keys to change as well, since 'Host' could be part of the
cache key if it's listed as Vary header.

Let's instead make a copy and mutate that instead.