Stop cloning the request headers
Since this is the only difference between haproxy and envoy I'm quite
sure that this is what's causing the weird errors in prod.

Turns out that Host is not a vary-header for any service and so it's
never part of any cache key, so it should be fine to mutate the request
headers.